requirements.txt
amqp==5.1.1
api==0.0.7
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==3.2
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
idna==3.4
kombu==5.2.4
Markdown==3.4.1
nose==1.3.7
prompt-toolkit==3.0.36
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
